requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.32.0
cryptography==39.0.0
defusedxml==0.7.1
dj-database-url==1.2.0
Django==3.2.7
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-allauth==0.52.0
django-auto-prefetch==1.4.0
django-cloudinary-storage==0.3.0
django-debug-toolbar==3.8.1
django-htmx==1.13.0
django-phonenumber-field==7.0.1
django-session-timeout==0.1.0
django-user-sessions==2.0.0
idna==3.4
oauthlib==3.2.2
Pillow==9.4.0
ply==3.11
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stone==3.3.1
tzdata==2022.7
urllib3==1.26.13
whitenoise==6.3.0
